feat: reorganização dos arquivos #45
