+++ backend/api/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import Funcionario, Material
+
+models_list = [Material, Funcionario]
+admin.site.register(models_list)
