+++ backend/api/models.py
@@ -14,3 +14,11 @@ class Funcionario(models.Model):
 
     def __str__(self):
         return self.nome
+    
+class Material(models.Model):
+    cod = models.AutoField(primary_key=True)
+    descricao = models.CharField(max_length=250)
+    quantEst = models.IntegerField()
+
+    def __str__(self):
+        return self.descricao
