+++ backend/backend/settings.py
@@ -39,8 +39,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     "rest_framework",
     "corsheaders",
-    "material.apps.MaterialConfig",
-    "funcionario.apps.FuncionarioConfig",
+    "api.apps.ApiConfig",
 ]
 
 MIDDLEWARE = [
