--- backend/material/serializers.py
@@ -1,10 +1,0 @@
-from rest_framework import serializers
-from .models import Material
-
-
-class MaterialSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Material
-        fields = ('cod',
-                  'descricao',
-                  'quantEst')
